+++ packages/analytics-js-plugins/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.11.1](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.11.0...@rudderstack/analytics-js-plugins@3.11.1) (2025-07-30)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-integrations` updated to version `3.16.0`
 ## [3.11.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.10.2...@rudderstack/analytics-js-plugins@3.11.0) (2025-07-21)
 
 ### Dependency Updates